+++ services/app-api/src/domain-models/contractAndRates/contractTypes.ts
@@ -1,78 +1,5 @@
 import { z } from 'zod'
-import { rateRevisionSchema } from './rateTypes'
-import { updateInfoSchema } from './updateInfoType'
-import {
-    contractExecutionStatusSchema,
-    contractTypeSchema,
-    federalAuthoritySchema,
-    populationCoveredSchema,
-    stateContactSchema,
-    submissionTypeSchema,
-} from '../../../../app-web/src/common-code/proto/healthPlanFormDataProto/zodSchemas'
-
-const documentSchema = z.object({
-    name: z.string(),
-    s3URL: z.string(),
-    sha256: z.string().optional(),
-})
-
-const managedCareEntitiesSchema = z.union([
-    z.literal('MCO'),
-    z.literal('PIHP'),
-    z.literal('PAHP'),
-    z.literal('PCCM'),
-])
-
-const contractFormDataSchema = z.object({
-    programIDs: z.array(z.string()),
-    populationCovered: populationCoveredSchema.optional(),
-    submissionType: submissionTypeSchema,
-    riskBasedContract: z.boolean().optional(),
-    submissionDescription: z.string(),
-    stateContacts: z.array(stateContactSchema),
-    supportingDocuments: z.array(documentSchema),
-    contractType: contractTypeSchema,
-    contractExecutionStatus: contractExecutionStatusSchema.optional(),
-    contractDocuments: z.array(documentSchema),
-    contractDateStart: z.date().optional(),
-    contractDateEnd: z.date().optional(),
-    managedCareEntities: z.array(managedCareEntitiesSchema),
-    federalAuthorities: z.array(federalAuthoritySchema),
-    inLieuServicesAndSettings: z.boolean().optional(),
-    modifiedBenefitsProvided: z.boolean().optional(),
-    modifiedGeoAreaServed: z.boolean().optional(),
-    modifiedMedicaidBeneficiaries: z.boolean().optional(),
-    modifiedRiskSharingStrategy: z.boolean().optional(),
-    modifiedIncentiveArrangements: z.boolean().optional(),
-    modifiedWitholdAgreements: z.boolean().optional(),
-    modifiedStateDirectedPayments: z.boolean().optional(),
-    modifiedPassThroughPayments: z.boolean().optional(),
-    modifiedPaymentsForMentalDiseaseInstitutions: z.boolean().optional(),
-    modifiedMedicalLossRatioStandards: z.boolean().optional(),
-    modifiedOtherFinancialPaymentIncentive: z.boolean().optional(),
-    modifiedEnrollmentProcess: z.boolean().optional(),
-    modifiedGrevienceAndAppeal: z.boolean().optional(),
-    modifiedNetworkAdequacyStandards: z.boolean().optional(),
-    modifiedLengthOfContract: z.boolean().optional(),
-    modifiedNonRiskPaymentArrangements: z.boolean().optional(),
-})
-
-const contractRevisionSchema = z.object({
-    id: z.string().uuid(),
-    submitInfo: updateInfoSchema.optional(),
-    unlockInfo: updateInfoSchema.optional(),
-    createdAt: z.date(),
-    updatedAt: z.date(),
-    formData: contractFormDataSchema,
-})
-
-// ContractRevision has all the information in a single submission of this contract.
-// If a revision has been submitted it will have submitInfo (otherwise it will be a draft)
-// if a revision was unlocked, it will have unlock info, otherwise it was an initial submission
-// The set of rateRevisions hold exactly what rate data was present at the time this contract was submitted.
-const contractRevisionWithRatesSchema = contractRevisionSchema.extend({
-    rateRevisions: z.array(rateRevisionSchema),
-})
+import { contractRevisionWithRatesSchema } from './revisionTypes'
 
 // Contract represents the contract specific information in a submission package
 // All that data is contained in revisions, each revision represents the data in a single submission
@@ -94,22 +21,7 @@ const draftContractSchema = contractSchema.extend({
 })
 
 type ContractType = z.infer<typeof contractSchema>
-type ContractRevisionType = z.infer<typeof contractRevisionSchema>
-type ContractRevisionWithRatesType = z.infer<
-    typeof contractRevisionWithRatesSchema
->
-type ContractFormDataType = z.infer<typeof contractFormDataSchema>
 
-export {
-    contractRevisionSchema,
-    contractRevisionWithRatesSchema,
-    draftContractSchema,
-    contractSchema,
-}
+export { contractRevisionWithRatesSchema, draftContractSchema, contractSchema }
 
-export type {
-    ContractType,
-    ContractRevisionType,
-    ContractRevisionWithRatesType,
-    ContractFormDataType,
-}
+export type { ContractType }
