+++ services/app-api/src/domain-models/contractAndRates/index.ts
@@ -1,29 +1,27 @@
+export { rateSchema, draftRateSchema } from './rateTypes'
+
+export type { RateType } from './rateTypes'
+
+export { contractSchema, draftContractSchema } from './contractTypes'
+
+export { contractFormDataSchema, rateFormDataSchema } from './formDataTypes'
+
 export {
-    rateSchema,
-    rateRevisionSchema,
     rateRevisionWithContractsSchema,
-    draftRateSchema,
-} from './rateTypes'
-
-export type {
-    RateType,
-    RateRevisionType,
-    RateRevisionWithContractsType,
-    RateFormDataType,
-} from './rateTypes'
-
-export {
-    contractSchema,
-    contractRevisionSchema,
     contractRevisionWithRatesSchema,
-    draftContractSchema,
-} from './contractTypes'
+    contractRevisionSchema,
+    rateRevisionSchema,
+} from './revisionTypes'
 
-export type {
-    ContractType,
-    ContractRevisionType,
-    ContractRevisionWithRatesType,
-    ContractFormDataType,
-} from './contractTypes'
+export type { ContractType } from './contractTypes'
 
 export type { ContractStatusType, UpdateInfoType } from './updateInfoType'
+
+export type { ContractFormDataType, RateFormDataType } from './formDataTypes'
+
+export type {
+    ContractRevisionType,
+    RateRevisionType,
+    RateRevisionWithContractsType,
+    ContractRevisionWithRatesType,
+} from './revisionTypes'
