+++ services/app-api/src/domain-models/contractAndRates/rateTypes.ts
@@ -1,115 +1,8 @@
 import { z } from 'zod'
 import {
-    actuaryCommunicationTypeSchema,
-    actuaryContactSchema,
-    rateCapitationTypeSchema,
-    rateTypeSchema,
-    sharedRateCertDisplay,
-    submissionDocumentSchema,
-    contractExecutionStatusSchema,
-    contractTypeSchema,
-    federalAuthoritySchema,
-    populationCoveredSchema,
-    stateContactSchema,
-    submissionTypeSchema,
-} from '../../../../app-web/src/common-code/proto/healthPlanFormDataProto/zodSchemas'
-import { updateInfoSchema } from './updateInfoType'
-
-const documentSchema = z.object({
-    name: z.string(),
-    s3URL: z.string(),
-    sha256: z.string().optional(),
-})
-
-const managedCareEntitiesSchema = z.union([
-    z.literal('MCO'),
-    z.literal('PIHP'),
-    z.literal('PAHP'),
-    z.literal('PCCM'),
-])
-
-// hacky workaround - importing from contract types is causing circular imports and this error
-// https://github.com/colinhacks/zod/issues/643#issuecomment-1015277287
-// we should resolve this issue to avoid duplicating efforts but hardcoding it for now to make progress
-
-const _contractFormDataSchema = z.object({
-    programIDs: z.array(z.string()),
-    populationCovered: populationCoveredSchema.optional(),
-    submissionType: submissionTypeSchema,
-    riskBasedContract: z.boolean().optional(),
-    submissionDescription: z.string(),
-    stateContacts: z.array(stateContactSchema),
-    supportingDocuments: z.array(documentSchema),
-    contractType: contractTypeSchema,
-    contractExecutionStatus: contractExecutionStatusSchema.optional(),
-    contractDocuments: z.array(documentSchema),
-    contractDateStart: z.date().optional(),
-    contractDateEnd: z.date().optional(),
-    managedCareEntities: z.array(managedCareEntitiesSchema),
-    federalAuthorities: z.array(federalAuthoritySchema),
-    inLieuServicesAndSettings: z.boolean().optional(),
-    modifiedBenefitsProvided: z.boolean().optional(),
-    modifiedGeoAreaServed: z.boolean().optional(),
-    modifiedMedicaidBeneficiaries: z.boolean().optional(),
-    modifiedRiskSharingStrategy: z.boolean().optional(),
-    modifiedIncentiveArrangements: z.boolean().optional(),
-    modifiedWitholdAgreements: z.boolean().optional(),
-    modifiedStateDirectedPayments: z.boolean().optional(),
-    modifiedPassThroughPayments: z.boolean().optional(),
-    modifiedPaymentsForMentalDiseaseInstitutions: z.boolean().optional(),
-    modifiedMedicalLossRatioStandards: z.boolean().optional(),
-    modifiedOtherFinancialPaymentIncentive: z.boolean().optional(),
-    modifiedEnrollmentProcess: z.boolean().optional(),
-    modifiedGrevienceAndAppeal: z.boolean().optional(),
-    modifiedNetworkAdequacyStandards: z.boolean().optional(),
-    modifiedLengthOfContract: z.boolean().optional(),
-    modifiedNonRiskPaymentArrangements: z.boolean().optional(),
-})
-// importing here bc of circular imports error
-const _contractRevisionSchema = z.object({
-    id: z.string().uuid(),
-    submitInfo: updateInfoSchema.optional(),
-    unlockInfo: updateInfoSchema.optional(),
-    createdAt: z.date(),
-    updatedAt: z.date(),
-    formData: _contractFormDataSchema,
-})
-
-// The rate form data  is the form filled out by state users submitting rates for review
-const rateFormDataSchema = z.object({
-    id: z.string().optional(),
-    rateType: rateTypeSchema.optional(),
-    rateCapitationType: rateCapitationTypeSchema.optional(),
-    rateDocuments: z.array(submissionDocumentSchema).optional(),
-    supportingDocuments: z.array(submissionDocumentSchema).optional(),
-    rateDateStart: z.date().optional(),
-    rateDateEnd: z.date().optional(),
-    rateDateCertified: z.date().optional(),
-    amendmentEffectiveDateStart: z.date().optional(),
-    amendmentEffectiveDateEnd: z.date().optional(),
-    rateProgramIDs: z.array(z.string()).optional(),
-    rateCertificationName: z.string().optional(),
-    certifyingActuaryContacts: z.array(actuaryContactSchema).optional(),
-    addtlActuaryContacts: z.array(actuaryContactSchema).optional(),
-    actuaryCommunicationPreference: actuaryCommunicationTypeSchema.optional(),
-    packagesWithSharedRateCerts: z.array(sharedRateCertDisplay).optional(),
-})
-
-// A Rate represents all the data associated with a single rate certification over time.
-// The rate's revisions field hold the array of rate revisions that show change history of specific rate data
-// The first revision (array index 0) is the current revision
-const rateRevisionSchema = z.object({
-    id: z.string().uuid(),
-    submitInfo: updateInfoSchema.optional(),
-    unlockInfo: updateInfoSchema.optional(),
-    createdAt: z.date(),
-    updatedAt: z.date(),
-    formData: rateFormDataSchema,
-})
-
-const rateRevisionWithContractsSchema = rateRevisionSchema.extend({
-    contractRevisions: z.array(_contractRevisionSchema),
-})
+    rateRevisionWithContractsSchema,
+    rateRevisionSchema,
+} from './revisionTypes'
 
 const rateSchema = z.object({
     id: z.string().uuid(),
@@ -127,11 +20,6 @@ const draftRateSchema = rateSchema.extend({
     revisions: z.array(rateRevisionWithContractsSchema).min(1),
 })
 type RateType = z.infer<typeof rateSchema>
-type RateRevisionType = z.infer<typeof rateRevisionSchema>
-type RateRevisionWithContractsType = z.infer<
-    typeof rateRevisionWithContractsSchema
->
-type RateFormDataType = z.infer<typeof rateFormDataSchema>
 
 export {
     rateRevisionSchema,
@@ -140,9 +28,4 @@ export {
     rateSchema,
 }
 
-export type {
-    RateType,
-    RateRevisionType,
-    RateRevisionWithContractsType,
-    RateFormDataType,
-}
+export type { RateType }
